Made connection to DB persistent
