+++ golang/main.go
@@ -21,6 +21,44 @@ func main() {
 	// Initializing router
 	router := mux.NewRouter()
 
+	var client *mongo.Client
+	client = nil
+	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
+	defer cancel()
+
+	// helpers will calls to the actual handlers
+	RegisterUserHandler := func(w http.ResponseWriter, r *http.Request) {
+		if client == nil {
+			// Secret URI
+			uri := os.Getenv("CFTMPR_ATLAS_URI")
+
+			client, err = mongo.Connect(ctx, options.Client().ApplyURI(uri))
+			if err != nil {
+				log.Fatal(err)
+			}
+		}
+		entry.RegisterUser(w, r, client)
+	}
+
+	LoginUserHandler := func(w http.ResponseWriter, r *http.Request) {
+		if client == nil {
+			// Secret URI
+			uri := os.Getenv("CFTMPR_ATLAS_URI")
+
+			client, err = mongo.Connect(ctx, options.Client().ApplyURI(uri))
+			if err != nil {
+				log.Fatal(err)
+			}
+		}
+		entry.LoginUser(w, r, client)
+	}
+
+	// Add API routes here
+
+	// User entry routes
+	router.HandleFunc("/api/register", RegisterUserHandler).Methods("POST") // registers new user
+	router.HandleFunc("/api/login", LoginUserHandler).Methods("POST")       // logins a user
+
 	// Add API routes here
 
 	// User entry routes
@@ -30,4 +68,6 @@ func main() {
 	// Start server
 	fmt.Println("Starting at port 8000...")
 	log.Fatal(http.ListenAndServe(":8000", router))
+
+	client.Disconnect(ctx)
 }
